Anjani: Database: split class into it's own file
