+++ anjani/util/db/errors.py
@@ -1,0 +1,1 @@
+from pymongo.errors import InvalidOperation, OperationFailure, PyMongoError
